requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
smartmoneyconcepts>=0.0.14
vectorbt>=0.26.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.30.0
schedule>=1.2.0
pytz>=2024.1
